- added stock name in backend - added support for fractional shares

+++ src/main/java/io/stock/portfolio/backend/controller/model/DividendResponse.java
@@ -20,6 +20,6 @@ public class DividendResponse {
     private BigDecimal euroNettoAmount;
     private LocalDateTime paymentDate;
     private LocalDateTime exDate;
-    private Integer shareAmount;
+    private BigDecimal shareAmount;
     private BigDecimal amountPerShare;
 }
